fix: address CodeRabbit review comments
- chapter_progress_utils: dedup using full double instead of toInt()
- manga_detail_page: replace hardcoded strings with l10n keys
- manga_detail_page: remove debugPrint from build()
- reading_progress_provider: optimistic state update before _save
- chapter_batch: copyWithFilteredItems also filters placeholders
- chapter_batch_list: plural batch range header, stable keys
- chapter_batch_list: pass manualThreshold to placeholder filter
- progress_jump_dialog: use jumpToChapterInvalid for empty input
- manga_detail_page_test: wire progressOverrides in pumpPage
- manga_detail_page_test: add 4th state (hasTotal, no MD chapters) test

+++ lib/features/library/domain/chapter_progress_utils.dart
@@ -42,16 +42,17 @@ List<Chapter> organizeChapters(
   bool descending = false,
   Set<String>? readChapterIds,
 }) {
-  // Deduplicate by chapter number keeping the first occurrence.
+  // Deduplicate by full chapter number keeping the first occurrence.
   // MangaDex sometimes returns multiple entries for the same number
-  // (different scanlators, multiple uploads).
-  final Set<int> seenNumbers = <int>{};
+  // (different scanlators, multiple uploads). Uses the full double so
+  // decimal chapters like 12.0 and 12.5 are both retained.
+  final Set<double> seenNumbers = <double>{};
   final List<Chapter> deduped = <Chapter>[];
   for (final chapter in chapters) {
     final num? n = chapter.number;
     if (n != null) {
-      if (seenNumbers.contains(n.toInt())) continue;
-      seenNumbers.add(n.toInt());
+      if (seenNumbers.contains(n.toDouble())) continue;
+      seenNumbers.add(n.toDouble());
     }
     deduped.add(chapter);
   }
